refactor sliced solvers by creating a dedicated directory
